Shaastra Login and Registration v1.0
The structure of the login part of the website with shaastra login-register along with Facebook
is complete. Since there are two ways for a user to register now, the registration hasn't yet been completed.
Both the methods should lead to the same register page and this unified registration page has to be given its functionality