Search bar completed.
JQuery has been implemented for a better search UI.
The earlier version displayed a dropdown box containing a list of hyperlinks. The jQuery based version has the facility to highlight search results, and navigate through them using arrow keys.
Also, now all tags are converted to lowercase while, adding, editing and searching, to bring about uniformity, and to ensure that results are found, irrespective of the case in which the query is entered.
Selecting a search result redirects the user to the appropriate page, containing the details of the particular event.